Added 7 new Java files
1. Added 7 new files in a new package named
org.cloudsim.examples.research

Files Added:

- CloudCreator.java
- CloudletExecutionTimeScenario.java
- CloudletTest.java
- CloudParameterGenerator.java
- CloudSimCloudletScenario.java
- CloudSimScenario.java
- TimeCalculator.java

2. The class CloudletsTableBuilder was slightly modified from the
original to not round to a double values for start, finish and execution
times to an integer (the roundTime method isn't used). So double values
for those times are printed in the simulation in the table. This change
affects all runs in the simulation

3. A new folder named Test_Data was added which contains CSV files for
running certain Java files in the research package.